codex: address follow-up review on PR #26840
