+++ codex-rs/utils/environment-id/BUILD.bazel
@@ -1,0 +1,6 @@
+load("//:defs.bzl", "codex_rust_crate")
+
+codex_rust_crate(
+    name = "environment-id",
+    crate_name = "codex_utils_environment_id",
+)
